+++ src/main/java/artifacts/item/curio/WhoopeeCushionItem.java
@@ -1,9 +1,0 @@
-package artifacts.item.curio;
-
-public class WhoopeeCushionItem extends TrinketArtifactItem {
-	@Override
-	public boolean canWearInSlot(String group, String slot) {
-		// Wear in any slot
-		return true;
-	}
-}