+++ src/main/java/artifacts2/mixin/extensions/DefaultedRegistryExtensions.java
@@ -1,0 +1,9 @@
+package artifacts2.mixin.extensions;
+
+import java.util.Optional;
+import net.minecraft.resources.ResourceLocation;
+
+public interface DefaultedRegistryExtensions<T> {
+
+	Optional<ResourceLocation> artifacts$getIdOrEmpty(T entry);
+}